requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
ctranslate2==4.4.0
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
faster-whisper==1.1.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.1.1
huggingface-hub==0.28.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.12.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnxruntime==1.20.1
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primePy==1.3
propcache==0.2.1
protobuf==5.29.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.2
SQLAlchemy==2.0.38
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.3.1+cu118
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.3.1+cu118
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.48.3
triton==2.3.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whisperx @ git+https://github.com/m-bain/whisperx.git@44e8bf5bb658ad6e1a80585b987fae2d48b9662c
yarl==1.18.3
